move ui progress to focus component
