+++ app/components/header/Header.jsx
@@ -1,4 +1,3 @@
-
 import { useDispatch, useSelector } from "react-redux";
 import "./header.css";
 import {
